+++ client/src/scss/Login.scss
@@ -25,7 +25,7 @@
     cursor: pointer;
     padding: 1rem;
     &:hover {
-      background: slategray;
+      background: white;
     }
     &.active {
       background: black;
